object Selected_3
Index Color값을 이용하여 Object 선택하는 기능 완성

+++ 14_part4/Examples/AppBase.h
@@ -209,7 +209,7 @@ class AppBase {
 
 
     // 거울이 아닌 물체들의 리스트 (for문으로 그리기 위함)
-    map<float, shared_ptr<Model>> m_objects;
+    map<int, shared_ptr<Model>> m_objects;
     vector<shared_ptr<Model>> m_basicList;
     vector<shared_ptr<Model>> m_pbrList;
     vector<shared_ptr<class Character>> m_characters;
